+++ HISTORY.md
@@ -1,13 +1,13 @@
 # History
 
-## 2.3.2 -> 3.0.0-alpha
+## 2.3.2 -> 3.0.0-alpha.1
 
 **Changed**
-- `&render.Config` -> `&rest.Config` . All configs that has relation only to the html/template are removed from there.
+- `&render.Config` -> `&iris.RestConfig` . All related to the html/template are removed from there.
 
 **Added**
 
-- Switch template engines via `IrisConfig`. Currently, Standar is 'html/template'. Pongo is 'flosch/pongo2`. Refer to the Book, which is updated too.
+- Switch template engines via `IrisConfig`. Currently, Standar is 'html/template'. Pongo is 'flosch/pongo2`. Refer to the Book, which is updated too, [read here](https://kataras.gitbooks.io/iris/content/render.html).
 
 
 ## 2.2.4 -> 2.3.0